+++ electron/preload.js
@@ -5,3 +5,7 @@ contextBridge.exposeInMainWorld("electronHotkey", {
   get: () => ipcRenderer.invoke("search-hotkey:get"),
   set: (value) => ipcRenderer.invoke("search-hotkey:set", value),
 });
+
+contextBridge.exposeInMainWorld("electronDesktop", {
+  startMicrosoftLogin: () => ipcRenderer.invoke("desktop-auth:start"),
+});